Transmit order
Options to show the NPC transmit order as overhead text above the NPC or in the right-click menu. Transmit order is the order in which the NPC was seen by the client. The first observed NPC will have number 0. A recently respawned NPC will have a high number. This changes every tick as NPCs move in and out of the scene or respawn/despawn. Ties are resolved by iterating 8x8 zones from south-west to north-east. Within 8x8 zones the ties are resolved based on which NPC entered the 8x8 zone first.